Werkzeuge einrichten: pyproject, ruff, mypy, pytest
Bisher gab es keine Werkzeug-Konfiguration - kein pyproject.toml, keinen Linter,
keinen Type-Checker, kein Packaging. Das ist jetzt da, und es hat sofort echte
Maengel gefunden.

ruff meldete zunaechst 879 Verstoesse. 628 davon waren Fehlalarme: RUF001/002/003
("ambiguous unicode") trifft in einem deutsch kommentierten Repo die Umlaute und die
typografischen Zeichen. Die drei Regeln sind mit Begruendung aus. Von den restlichen
251 sind alle behoben:

- 47 ungenutzte Importe - Rueckstand aus den generierten Import-Koepfen der
  Aufteilungs-Skripte
- 77 unsortierte Import-Bloecke
- 24 %-Formatierungen zu f-strings, 18 redundante int(round(...))-Wrapper
- 13-mal die Variable l (verwechselbar mit 1 und I) sinnvoll benannt: lim fuer ein
  Limit, lbl fuer ein Label, left/top/right/bottom fuer Rechteck-Koordinaten
- 7-mal zip() ohne strict, 4 lambda-Zuweisungen, ungenutzte Variablen, assert False

Zwei Funde, die mehr als Kosmetik waren:

Meine eigenen Architektur-Tests liessen je Lauf ueber hundert Dateihandles offen
(open(...).read() ohne with). Sichtbar wurde das erst durch filterwarnings = "error"
in der pytest-Konfiguration - pytest meldet es als unraisable ResourceWarning. Jetzt
liest eine _src()-Funktion mit Context-Manager, und SIM115 ist als ruff-Regel an,
damit es nicht wiederkommt. Das Log-Handle in ops/log.py bleibt dagegen bewusst offen
(faulthandler schreibt dorthin) und ist als Ausnahme markiert.

mypy fand 6 echte Stellen: viermal das Muster "try: dll = WinDLL(...) except: dll =
None" ohne Optional-Typ, zweimal einen leeren Container ohne ableitbaren Inhalt
(_resize_hooks, _token_cache). Alle typisiert.

Die CI prueft jetzt vier Dinge statt einem: compileall, ruff, mypy und BEIDE
Testwege - pytest und den eigenen Runner. Der Runner bleibt, weil das Repo ohne
Test-Abhaengigkeit lauffaehig sein soll; pytest braucht dank pythonpath in
pyproject.toml keinen sys.path-Hack mehr.

Notiert statt uebersehen: Pillow entfernt Image.getdata() in Version 14 (Oktober
2027). Die Warnung ist gefiltert, die Begruendung steht dabei.

Geprueft: ruff clean, mypy clean (35 Dateien), 149/149 in beiden Testwegen, Panel neu
gestartet.

+++ deck/platform/dpi.py
@@ -120,7 +120,7 @@ def ui():
 def px(v):
     """Design-Einheit (Mass bei 100 %) -> Geraetepixel, ganzzahlig gerundet.
     Fuer feste Masse ausserhalb des Kachel-Renderers (Leistenhoehe, Griffdicke …)."""
-    return int(round(v * _ui))
+    return round(v * _ui)
 
 
 def fpx(v):
@@ -153,5 +153,5 @@ def fontpx(pt, scale=1.0, family="Segoe UI", weight=None):
     Bei 150 % und Zoom 1.0 wird aus 10 pt: 10 * 1.333 * 1.5 = 20 Pixel – also
     dieselbe optische Groesse wie vor der Umstellung, nur echt gezeichnet statt
     hochgerechnet."""
-    size = max(1, int(round(pt * PT_PX * scale)))
+    size = max(1, round(pt * PT_PX * scale))
     return (family, -size, weight) if weight else (family, -size)
